conf: remove all dead code and bullshit tests
**Removed dead code (502 lines):**
- lib/configschema/legacy_jsonschema_parser.go (410 lines) - no longer needed
- conf/pkg/schemas/loader.go (215 lines) - SchemaLoader never used
- conf/pkg/schemas/loader_test.go - tests for unused SchemaLoader
- conf/pkg/schemas/mise.go (76 lines) - TOML schema parser, never used
- conf/pkg/schemas/mise_test.go - tests for unused mise.go
- All embed re-export files (30 lines):
  - claude_embed.go, embed.go, mise_json_embed.go,
    mise_embed.go, starship_embed.go
  - Only used by deleted loader.go

**Removed bullshit tests:**
- TestHelperFunctions - tested internal helper functions (not our responsibility)
- TestJJSchemaEmbedded - tested Go's embed directive (not our responsibility)
- TestMiseSchemaEmbedded - tested Go's embed directive (not our responsibility)

**Added compile.go (84 lines):**
- Extracted CompileJJSchema, CompileClaudeSchema, CompileMiseSchema, CompileStarshipSchema
- These are the only functions from loader.go that were actually used
- Now directly call lib/configschema schema accessors instead of re-exporting

**Final state - conf/pkg/schemas (1,175 lines total):**
✅ compile.go (84 lines) - Schema compilation functions
✅ 4 thin parser wrappers (~45 lines each):
   - jj_parser.go, claude_parser.go, mise_json_parser.go, starship_parser.go
✅ Test files for parsers + jsonschema_parser_test.go

All tests pass (45 tests).
