--- conf/pkg/schemas/claude_embed.go
@@ -1,6 +1,0 @@
-package schemas
-
-import "github.com/neongreen/mono/lib/configschema"
-
-// ClaudeSchema contains the embedded Claude Code configuration schema
-var ClaudeSchema string = configschema.ClaudeSchema()
