--- conf/pkg/schemas/embed.go
@@ -1,6 +1,0 @@
-package schemas
-
-import "github.com/neongreen/mono/lib/configschema"
-
-// JJSchema contains the embedded jj configuration schema (latest version)
-var JJSchema string = configschema.JJSchemaLatest()
