--- conf/pkg/schemas/mise_embed.go
@@ -1,8 +1,0 @@
-package schemas
-
-import _ "embed"
-
-// MiseSchemaData contains the embedded mise configuration schema
-//
-//go:embed mise.schema
-var MiseSchemaData string
