--- conf/pkg/schemas/mise_json_embed.go
@@ -1,6 +1,0 @@
-package schemas
-
-import "github.com/neongreen/mono/lib/configschema"
-
-// MiseJSONSchema contains the embedded Mise configuration JSON schema
-var MiseJSONSchema string = configschema.MiseSchema()
