--- conf/pkg/schemas/starship_embed.go
@@ -1,6 +1,0 @@
-package schemas
-
-import "github.com/neongreen/mono/lib/configschema"
-
-// StarshipSchema contains the embedded starship configuration schema
-var StarshipSchema string = configschema.StarshipSchema()
